packages/cacl/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "cacl"
version = "0.1.7"
description = "Stop fighting auth in FastAPI: JWT access/refresh, DB-backed blacklist & rotation, cookie or Bearer modes."
readme = "README_PYPI.md"
requires-python = ">=3.11"
license = "MIT"
authors = [{ name = "Anton Serebriakov" }]
keywords = [
  "fastapi",
  "authentication",
  "jwt",
  "security",
  "access-token",
  "refresh-token",
  "cookie-auth",
  "bearer-token",
  "token-blacklist",
  "auth",
  "tokens",
  "refresh-rotation",
  "sqlalchemy",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Framework :: FastAPI",
    "Topic :: Security",
    "Topic :: Internet :: WWW/HTTP :: Session",
    "Topic :: Software Development :: Libraries",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "fastapi>=0.100.0",
    "sqlalchemy>=2.0.0",
    "python-jose>=3.3.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
postgres = [
    "asyncpg>=0.29.0",
    "greenlet>=3.0.0",
]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[tool.setuptools]
packages = ["cacl", "cacl.models", "cacl.services", "cacl.utils"]

[tool.setuptools.package-dir]
"" = "."

# Exclude tests from sdist
[tool.setuptools.exclude-package-data]
"*" = ["tests*"]
